1. 临时时提交,同步.

+++ xplat-meta-oqsengine-devops/src/main/java/com/xforceplus/ultraman/oqsengine/devops/rebuild/model/IDevOpsTaskInfo.java
@@ -60,7 +60,7 @@ public interface IDevOpsTaskInfo {
 
     IEntityClass getEntityClass();
 
-    int getBatchSize();
+    long getBatchSize();
 
     int getFinishSize();
 
@@ -76,7 +76,7 @@ public interface IDevOpsTaskInfo {
 
     void setOffsetSnapShot(OffsetSnapShot offsetSnapShot);
 
-    void setBatchSize(int size);
+    void setBatchSize(long size);
 
     int failedRecovers();
